Initial steps for Feature component
